+++ src/main/kotlin/io/kjson/spring/test/JSONMockServerDSL.kt
@@ -55,45 +55,97 @@ class JSONMockServerDSL private constructor() : ResponseCreator {
 
     private lateinit var request: MockClientHttpRequest
 
+    private var response: Response? = null
+        set(newResponse) {
+            if (field != null && newResponse != null)
+                throw AssertionError("Response already set")
+            field = newResponse
+        }
+
+    private fun reset() {
+        response = null
+    }
+
+    /**
+     * Match the request URI using a string.
+     */
     fun requestTo(expectedURI: String) {
         requestTo(URI(expectedURI))
     }
 
+    /**
+     * Match the URI against another URI.
+     */
     fun requestTo(expectedURI: URI) {
         val uri = request.uri
         if (!uri.equalIgnoringQuery(expectedURI))
             fail("Request URI doesn't match; expected $expectedURI, was $uri")
     }
 
+    /**
+     * Match the URI using a lambda.
+     */
     fun requestTo(test: (String) -> Boolean) {
         if (!test(request.uri.toString()))
             fail("Request URI doesn't match; was ${request.uri}")
     }
 
+    /**
+     * Match the URI using a [Matcher].
+     */
     @Deprecated("The use of Matcher will be removed in a future version", ReplaceWith("requestTo { test(it) }"))
     fun requestTo(matcher: Matcher<in String>) {
         MatcherAssert.assertThat("Request URI", request.uri.toString(), matcher)
     }
 
+    /**
+     * Match the method.
+     */
     fun method(method: HttpMethod) {
         if (request.method != method)
             fail("Request method incorrect; expected $method, was ${request.method}")
     }
 
+    /**
+     * Match a query parameter.  A query parameter with multiple values will be matched against a set of values.
+     */
     fun queryParam(name: String, vararg expectedValues: String) {
-        val queryEntries = request.uri.query?.split('&')?.filter { it.startsWith("$name=") }
+        val queryMap = JSONMockClientRequest.decodeQueryParams(request.uri)
+        val queryEntries = queryMap[name]
         if (queryEntries == null || queryEntries.isEmpty())
             throw AssertionError("Request query param [$name] not found")
         val n = expectedValues.size
-        if (queryEntries.size != n)
-            fail("Request query param [$name] number incorrect; expected $n, was ${queryEntries.size}")
+        val s = queryEntries.size
+        if (s != n) {
+            when (n) {
+                1 -> fail("Request query param [$name] incorrect; expected single param, was multiple ($s)")
+                else -> fail("Request query param [$name] number incorrect; expected $n, was $s")
+            }
+        }
         for (i in 0 until n) {
-            val queryValue = queryEntries[i].substringAfter('=')
-            if (expectedValues[i] != queryValue)
-                fail("Request query param [$name] incorrect; expected ${expectedValues[i]}, was $queryValue")
+            if (expectedValues[i] != queryEntries[i])
+                fail("Request query param [$name] incorrect; expected ${expectedValues[i]}, was ${queryEntries[i]}")
         }
     }
 
+    /**
+     * Match a query parameter using a lambda.
+     */
+    fun queryParam(name: String, test: (String) -> Boolean) {
+        val queryMap = JSONMockClientRequest.decodeQueryParams(request.uri)
+        val queryEntries = queryMap[name]
+        if (queryEntries == null || queryEntries.isEmpty())
+            throw AssertionError("Request query param [$name] not found")
+        if (queryEntries.size != 1)
+            fail("Request query param [$name] incorrect; expected single param, was multiple (${queryEntries.size})")
+        val param = queryEntries[0]
+        if (param == null || !test(param))
+            fail("Request query param [$name] incorrect")
+    }
+
+    /**
+     * Match a request header.  A header with multiple values will be matched against a set of values.
+     */
     fun header(name: String, vararg expectedValues: String) {
         val n = expectedValues.size
         val headers = getHeaders(name, n)
@@ -102,12 +154,18 @@ class JSONMockServerDSL private constructor() : ResponseCreator {
                 fail("Request header [$name] incorrect; expected ${expectedValues[i]}, was ${headers[i]}")
     }
 
+    /**
+     * Match a request header using a lambda.
+     */
     fun header(name: String, test: (String) -> Boolean) {
         val header = getHeaders(name, 1).first()
         if (!test(header))
             fail("Request header [$name] incorrect; was $header")
     }
 
+    /**
+     * Match a request header using a [Matcher].
+     */
     @Deprecated("The use of Matcher will be removed in a future version", ReplaceWith("header(name) { test(it) }"))
     fun header(name: String, vararg matchers: Matcher<in String?>) {
         val n = matchers.size
@@ -116,6 +174,10 @@ class JSONMockServerDSL private constructor() : ResponseCreator {
             MatcherAssert.assertThat("Request header [$name]", headers[i], matchers[i])
     }
 
+    /**
+     * Match the `Accept` header against a specified [MediaType].  An `Accept` header with multiple values will be
+     * considered to match if any of the entries is compatible with the expected type.
+     */
     fun accept(expectedMediaType: MediaType) {
         val header = getHeaders(HttpHeaders.ACCEPT, 1).first()
         header.split(',').map { it.trim() }.forEach {
@@ -125,16 +187,25 @@ class JSONMockServerDSL private constructor() : ResponseCreator {
         fail("Request [Accept] header incorrect; expected $expectedMediaType, was $header")
     }
 
+    /**
+     * Match the `Accept` header with `application/json`.
+     */
     fun acceptApplicationJSON() {
         accept(MediaType.APPLICATION_JSON)
     }
 
+    /**
+     * Match the `Content-Type` header against a specified [MediaType].
+     */
     fun contentType(expectedMediaType: MediaType) {
         val header = getHeaders(HttpHeaders.CONTENT_TYPE, 1).first()
         if (!checkMediaType(header, HttpHeaders.CONTENT_TYPE).isCompatibleWith(expectedMediaType))
             fail("Request [Content-Type] header incorrect; expected $expectedMediaType, was $header")
     }
 
+    /**
+     * Match the `Content-Type` header with `application/json`.
+     */
     fun contentTypeApplicationJSON() {
         contentType(MediaType.APPLICATION_JSON)
     }
@@ -142,72 +213,205 @@ class JSONMockServerDSL private constructor() : ResponseCreator {
     private fun getHeaders(name: String, expectedCount: Int): List<String> {
         val headers = request.headers[name] ?: throw AssertionError("Header [$name] not found")
         if (headers.size != expectedCount) {
-            if (expectedCount == 1)
-                fail("Request [$name] header; expected single header, was multiple (${headers.size})")
-            else
-                fail("Request [$name] header number incorrect; expected $expectedCount, was ${headers.size}")
+            when (expectedCount) {
+                1 -> fail("Request [$name] header; expected single header, was multiple (${headers.size})")
+                else -> fail("Request [$name] header number incorrect; expected $expectedCount, was ${headers.size}")
+            }
         }
         return headers
     }
 
+    /**
+     * Test that the nominated request header is not present in the request.
+     */
     fun headerDoesNotExist(name: String) {
         if (request.headers.containsKey(name))
             fail("Request [$name] header expected not to be present")
     }
 
+    /**
+     * Match the request body content against a string.
+     */
     fun requestContent(body: String) {
         if (request.bodyAsString != body)
             fail("Request body incorrect")
     }
 
+    /**
+     * Match the request body content using a lambda.
+     */
     fun requestContent(test: (String) -> Boolean) {
         if (!test(request.bodyAsString))
             fail("Request body incorrect")
     }
 
+    /**
+     * Match the request body using [JSONExpect] tests.
+     */
     fun requestJSON(tests: JSONExpect.() -> Unit) {
         contentTypeApplicationJSON()
         JSONExpect.expectJSON(request.bodyAsString, tests)
     }
 
-    private var response: Response? = null
-
+    /**
+     * Specify the response from the request as an object to be serialised to JSON.  The object will be created
+     * dynamically, and the lambda that creates it will have access to the request variables.
+     */
     fun respondJSON(
         status: HttpStatus = HttpStatus.OK,
         headers: HttpHeaders? = null,
-        block: MockClientHttpRequest.() -> Any?
+        lambda: JSONMockClientRequest.() -> Any?
     ) {
-        respondJSON(status, headers, request.block())
+        response = Response(
+            status = status,
+            headers = combineHeaders(headers, MediaType.APPLICATION_JSON),
+            lambda = { lambda().stringifyJSON(JSONTestConfig.config).toByteArray() },
+        )
     }
 
+    /**
+     * Specify the response from the request as a fixed object to be serialised to JSON.
+     */
     fun respondJSON(
         status: HttpStatus = HttpStatus.OK,
         headers: HttpHeaders? = null,
         result: Any?,
     ) {
-        val headers1 = HttpHeaders()
-        headers?.let { headers1.addAll(it) }
-        headers1.contentType = MediaType.APPLICATION_JSON
-        respond(status, headers1, result.stringifyJSON(JSONTestConfig.config))
+        response = Response(
+            status = status,
+            headers = combineHeaders(headers, MediaType.APPLICATION_JSON),
+            body = result.stringifyJSON(JSONTestConfig.config).toByteArray()
+        )
     }
 
+    /**
+     * Specify the response from the request as a fixed string.
+     */
     fun respond(
+        status: HttpStatus = HttpStatus.OK,
+        headers: HttpHeaders? = null,
+        contentType: MediaType? = null,
+        result: String? = null,
+    ) {
+        response = Response(
+            status = status,
+            headers = combineHeaders(headers, contentType),
+            body = result?.toByteArray(),
+        )
+    }
+
+    /**
+     * Specify the response from the request as a string.  The string will be created dynamically, and the lambda that
+     * creates it will have access to the request variables.
+     */
+    fun respond(
+        status: HttpStatus = HttpStatus.OK,
+        headers: HttpHeaders? = null,
+        contentType: MediaType? = null,
+        lambda: JSONMockClientRequest.() -> String?
+    ) {
+        response = Response(
+            status = status,
+            headers = combineHeaders(headers, contentType),
+            lambda = { lambda()?.toByteArray() },
+        )
+    }
+
+    /**
+     * Specify the response from the request as a fixed string, with the type `text/plain`.
+     */
+    fun respondTextPlain(
         status: HttpStatus = HttpStatus.OK,
         headers: HttpHeaders? = null,
         result: String? = null,
     ) {
-        response = Response(status, headers, result)
+        response = Response(
+            status = status,
+            headers = combineHeaders(headers, MediaType.TEXT_PLAIN),
+            body = result?.toByteArray(),
+        )
     }
 
+    /**
+     * Specify the response from the request as a string, with the type `text/plain`.  The string will be created
+     * dynamically, and the lambda that creates it will have access to the request variables.
+     */
+    fun respondTextPlain(
+        status: HttpStatus = HttpStatus.OK,
+        headers: HttpHeaders? = null,
+        lambda: JSONMockClientRequest.() -> String?
+    ) {
+        response = Response(
+            status = status,
+            headers = combineHeaders(headers, MediaType.TEXT_PLAIN),
+            lambda = { lambda()?.toByteArray() },
+        )
+    }
+
+    /**
+     * Specify the response from the request as a fixed byte array.
+     */
+    fun respondBytes(
+        status: HttpStatus = HttpStatus.OK,
+        headers: HttpHeaders? = null,
+        contentType: MediaType? = null,
+        result: ByteArray? = null,
+    ) {
+        response = Response(
+            status = status,
+            headers = combineHeaders(headers, contentType),
+            body = result,
+        )
+    }
+
+    /**
+     * Specify the response from the request as a fixed byte array.  The byte array will be created dynamically, and the
+     * lambda that creates it will have access to the request variables.
+     */
+    fun respondBytes(
+        status: HttpStatus = HttpStatus.OK,
+        headers: HttpHeaders? = null,
+        contentType: MediaType? = null,
+        lambda: JSONMockClientRequest.() -> ByteArray?
+    ) {
+        response = Response(
+            status = status,
+            headers = combineHeaders(headers, contentType),
+            lambda = lambda,
+        )
+    }
+
+    /**
+     * Create the [ClientHttpResponse] for this request, **if** the response has been set.  If not, `null` will be
+     * returned, allowing the response to be set using a chained `andRespond()` function.
+     */
     @Suppress("WRONG_NULLABILITY_FOR_JAVA_OVERRIDE")
     override fun createResponse(request: ClientHttpRequest?): ClientHttpResponse? {
         return response?.let { resp ->
-            val body = resp.body?.toByteArray() ?: ByteArray(0)
+            val body = resp.getValue(request as MockClientHttpRequest)
             MockClientHttpResponse(body, resp.status).apply { resp.headers?.let { headers.addAll(it) } }
         }
     }
 
-    data class Response(val status: HttpStatus, val headers: HttpHeaders?, val body: String?)
+    class Response(
+        val status: HttpStatus,
+        val headers: HttpHeaders?,
+        private val body: ByteArray? = null,
+        private val lambda: (JSONMockClientRequest.() -> ByteArray?)? = null
+    ) {
+
+        init {
+            if (body != null && lambda != null)
+                throw RuntimeException("Response body and lambda may not both be specified")
+        }
+
+        fun getValue(mockClientHttpRequest: MockClientHttpRequest): ByteArray {
+            body?.let { return it }
+            lambda?.let { return JSONMockClientRequest(mockClientHttpRequest).it() ?: ByteArray(0) }
+            return ByteArray(0)
+        }
+
+    }
 
     companion object {
 
@@ -227,6 +431,17 @@ class JSONMockServerDSL private constructor() : ResponseCreator {
                 !other.isOpaque && scheme == other.scheme && userInfo == other.userInfo && host == other.host &&
                         port == other.port && path == other.path
 
+        fun combineHeaders(headers: HttpHeaders?, contentType: MediaType?): HttpHeaders? = when {
+            contentType != null -> HttpHeaders().also { h ->
+                headers?.let { h.addAll(it) }
+                h.contentType = contentType
+            }
+            else -> headers
+        }
+
+        /**
+         * Establish a mock request.
+         */
         fun MockRestServiceServer.mock(
             expectedCount: ExpectedCount = ExpectedCount.once(),
             method: HttpMethod? = null,
@@ -237,6 +452,7 @@ class JSONMockServerDSL private constructor() : ResponseCreator {
             val responseActions = expect(expectedCount) { request ->
                 serverDSL.request = request as MockClientHttpRequest
                 serverDSL.apply {
+                    reset() // allow "respond" functions to set the response
                     method?.let { method(it) }
                     uri?.let { requestTo(it) }
                     block()
@@ -249,12 +465,18 @@ class JSONMockServerDSL private constructor() : ResponseCreator {
             return responseActions
         }
 
+        /**
+         * Establish a mock request with the method preset to GET.
+         */
         fun MockRestServiceServer.mockGet(
             expectedCount: ExpectedCount = ExpectedCount.once(),
             uri: URI? = null,
             block: JSONMockServerDSL.() -> Unit = {}
         ) = mock(expectedCount, HttpMethod.GET, uri, block)
 
+        /**
+         * Establish a mock request with the method preset to POST.
+         */
         fun MockRestServiceServer.mockPost(
             expectedCount: ExpectedCount = ExpectedCount.once(),
             uri: URI? = null,